policies

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,10 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Category;
 use App\Http\Controllers\Controller;
 use App\Services\Admin\CategoryService;
-use App\Http\Requests\Admin\StoreCategoryRequest;
-use App\Http\Requests\Admin\ImportCategoryRequest;
-use App\Http\Requests\Admin\UpdateCategoryRequest;
-use App\Http\Requests\Admin\BulkDeleteCategoryRequest;
+use App\Http\Requests\Admin\{StoreCategoryRequest, ImportCategoryRequest, UpdateCategoryRequest, BulkDeleteCategoryRequest};
 
 class CategoryController extends Controller
 {
@@ -21,6 +18,7 @@ class CategoryController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Category::class);
         $data['categories'] = Category::with('subcategories')->whereNull('parent_id')->orderby('name', 'asc')->get();
         return view('backend.category.create', $data);
     }
@@ -28,19 +26,19 @@ class CategoryController extends Controller
 
     public function store(StoreCategoryRequest $request, CategoryService $service)
     {
+        $this->authorize('create', Category::class);
         $service->create($request->validated());
         toastr()->success('Category created successfully');
         return redirect()->route('category.index');
     }
 
 
-    public function edit(string $id)
+    public function edit(Category $category)
     {
-        $category = Category::findOrFail($id);
-
+        $this->authorize('update', $category);
         $categories = Category::with('subcategories.subcategories')
             ->whereNull('parent_id')
-            ->where('id', '!=', $id)
+            ->where('id', '!=', $category->id)
             ->orderby('name', 'asc')
             ->get();
 
@@ -53,6 +51,7 @@ class CategoryController extends Controller
         Category $category,
         CategoryService $service
     ) {
+        $this->authorize('update', $category);
         $service->update($category, $request->validated());
 
         toastr()->success('Category updated successfully');
@@ -61,10 +60,9 @@ class CategoryController extends Controller
 
 
 
-    public function destroy($id, CategoryService $service)
+    public function destroy(Category $category, CategoryService $service)
     {
-        $category = Category::findOrFail($id);
-
+        $this->authorize('delete', $category);
         $service->delete($category);
 
         toastr()->success('Category Deleted Successfully');
@@ -74,8 +72,10 @@ class CategoryController extends Controller
 
     public function bulkDelete(
         BulkDeleteCategoryRequest $request,
-        CategoryService $service
+        CategoryService $service,
+        Category $category
     ) {
+        $this->authorize('delete', $category);
         $service->bulkDelete($request->getCategoryIds());
 
         toastr()->success('Categories deleted successfully');
